+++ mediation/AppLovin/source/plugin/Assets/GoogleMobileAds/Platforms/iOS/Mediation/AppLovin/AppLovinClient.cs
@@ -38,6 +38,20 @@ namespace GoogleMobileAds.iOS.Mediation.AppLovin
         {
             Externs.GADUMInitializeAppLovin();
         }
+
+        public void SetHasUserConsent(bool hasUserConsent)
+        {
+            string parameterString = (hasUserConsent == true ? "YES" : "NO");
+            MonoBehaviour.print("Calling '[ALPrivacySettings setHasUserConsent:]' with argument: " + parameterString);
+            Externs.GADUMAppLovinSetHasUserConsent (hasUserConsent);
+        }
+
+        public void SetIsAgeRestrictedUser(bool isAgeRestrictedUser)
+        {
+            string parameterString = (isAgeRestrictedUser == true ? "YES" : "NO");
+            MonoBehaviour.print("Calling '[ALPrivacySettings setIsAgeRestrictedUser:]' with argument: " + parameterString);
+            Externs.GADUMAppLovinSetIsAgeRestrictedUser (isAgeRestrictedUser);
+        }
     }
 }