+++ mediation/AppLovin/source/plugin/Assets/GoogleMobileAds/Platforms/iOS/Mediation/AppLovin/Externs.cs
@@ -24,6 +24,12 @@ namespace GoogleMobileAds.iOS.Mediation.AppLovin
     {
         [DllImport("__Internal")]
         internal static extern void GADUMInitializeAppLovin();
+
+        [DllImport("__Internal")]
+        internal static extern void GADUMAppLovinSetHasUserConsent(bool hasUserConsent);
+
+        [DllImport("__Internal")]
+        internal static extern void GADUMAppLovinSetIsAgeRestrictedUser(bool isAgeRestrictedUser);
     }
 }