+++ mediation/AdColony/CHANGELOG.md
@@ -1,7 +1,22 @@
 # AdColony Adapter plugin for Google Mobile Ads SDK for Unity 3D Changelog
 
-## 1.0.0
+## 1.0.3
+- Supports AdColony Android adapter version 3.3.5.1.
+- Supports AdColony iOS adapter version 3.3.5.0.
 
+## 1.0.2
+- Supports AdColony Android SDK version 3.3.4.
+- Supports AdColony iOS SDK version 3.3.4.
+- `SetTestMode()` from the `AdColonyMediationExtras` Builder class is now ignored for Android. Publishers can now request test ads from AdColony for Android by specifying a test device via `AddTestDevice()` from the `AdRequest` Builder class.
+- Added the following methods to the `AdColonyMediationExtras` Builder class:
+  * `SetGDPRRequired()`
+  * `SetGDPRConsentString()`
+
+## 1.0.1
+- Supports AdColony Android SDK version 3.3.0-unity.
+- Supports AdColony iOS SDK version 3.3.0.
+
+## 1.0.0
 - First release!
-- Supports Android adapter version 3.3.0.0.
-- Supports iOS adapter version 3.3.0.0.
+- Supports AdColony Android SDK version 3.3.0.
+- Supports AdColony iOS SDK version 3.3.0.