+++ mediation/AppLovin/CHANGELOG.md
@@ -1,5 +1,43 @@
 # AppLovin Adapter plugin for Google Mobile Ads SDK for Unity 3D Changelog
 
+## 3.1.1
+- Supports AppLovin Android adapter version 8.1.0.0.
+- Supports AppLovin iOS adapter version 5.1.1.0.
+
+## 3.1.0
+- Supports AppLovin Android adapter version 8.1.0.0.
+- Supports AppLovin iOS adapter version 5.1.0.0.
+
+## 3.0.3
+- Supports AppLovin Android adapter version 8.0.2.1.
+- Supports AppLovin iOS adapter version 5.0.2.0.
+
+## 3.0.2
+- Supports AppLovin Android SDK version 8.0.1.
+- Supports AppLovin iOS SDK version 5.0.1.
+
+## 3.0.1
+- Supports AppLovin Android SDK version 8.0.0.
+- Supports AppLovin iOS SDK version 5.0.1.
+- Added `AppLovin.SetIsAgeRestrictedUser()` method to indicate if the user is known to be in an age-restricted category.
+
+## 3.0.0
+- Supports AppLovin Android SDK version 8.0.0.
+- Supports AppLovin iOS SDK version 5.0.1.
+- Added `AppLovin.SetHasUserConsent()` method to forward user consent flag to the AppLovin SDK.
+
+## 2.0.0
+- Supports AppLovin Android SDK version 7.8.6.
+- Supports AppLovin iOS SDK version 5.0.1.
+
+## 1.2.1
+- Supports AppLovin Android SDK version 7.8.6.
+- Supports AppLovin iOS SDK version 4.8.3.
+
+## 1.2.0
+- Supports AppLovin Android SDK version 7.8.5.
+- Supports AppLovin iOS SDK version 4.8.3.
+
 ## 1.1.0
 - Supports AppLovin Android SDK version 7.7.0.
 - Supports AppLovin iOS SDK version 4.7.0.