+++ mediation/MyTarget/CHANGELOG.md
@@ -1,7 +1,39 @@
 # myTarget Adapter plugin for Google Mobile Ads SDK for Unity 3D Changelog
 
-## 1.0.0
+## 2.2.0
+- Supports myTarget Android adapter version 5.1.4.0.
+- Supports myTarget iOS adapter version 4.8.0.0.
 
+## 2.1.2
+- Supports myTarget Android adapter version 5.1.3.0.
+- Supports myTarget iOS adapter version 4.7.11.0.
+
+## 2.1.1
+- Supports myTarget Android adapter version 5.1.1.0.
+- Supports myTarget iOS adapter version 4.7.10.0.
+
+## 2.1.0
+- Supports myTarget Android SDK version 5.1.0.
+- Supports myTarget iOS SDK version 4.7.9.
+- Added the following methods:
+  * `MyTarget.SetUserConsent()` method to forward the user's consent status to the MyTarget SDK.
+  * `MyTarget.SetUserAgeRestricted` method that sets a flag to indicate if the user is known to be in an age-restricted category (i.e., under the age of 16).
+  * `MyTarget.IsUserConsent()` method to get the user's current consent status.
+  * `MyTarget.IsUserAgeRestricted()` method to get the flag indicating that the user is known to be in an age-restricted category (i.e., under the age of 16).
+
+## 2.0.1
+- Supports myTarget Android SDK version 5.0.2.
+- Supports myTarget iOS SDK version 4.7.8.
+
+## 2.0.0
+- Supports myTarget Android SDK version 5.0.0.
+- Supports myTarget iOS SDK version 4.7.7.
+
+## 1.1.0
+- Supports myTarget Android SDK version 4.7.2.
+- Supports myTarget iOS SDK version 4.7.6.
+
+## 1.0.0
 - First release!
-- Supports Android adapter version 4.6.27.0.
-- Supports iOS adapter version 4.6.24.0.
+- Supports myTarget Android SDK version 4.6.27.
+- Supports myTarget iOS SDK version 4.6.24.