+++ mediation/Tapjoy/CHANGELOG.md
@@ -1,5 +1,27 @@
 # Tapjoy Adapter plugin for Google Mobile Ads SDK for Unity 3D Changelog
 
+## 2.0.1
+
+- Supports Tapjoy Android adapter version 12.0.0.1.
+- Supports Tapjoy iOS adapter version 12.0.0.0.
+
+## 2.0.0
+
+- Supports Tapjoy Android adapter version 12.0.0.0.
+- Supports Tapjoy iOS adapter version 12.0.0.0.
+
+## 1.1.1
+
+- Supports Tapjoy Android SDK version 11.12.2.
+- Supports Tapjoy iOS SDK version 11.12.2.
+- Added `Tapjoy.SetUserConsent()` method to forward the user's consent status to the Tapjoy SDK in the form of a consent string.
+- Added `Tapjoy.SubjectToGDPR()` method to indicate if GDPR is applicable to the user or not.
+
+## 1.1.0
+
+- Supports Tapjoy Android SDK version 11.12.1.
+- Supports Tapjoy iOS SDK version 11.12.1.
+
 ## 1.0.0
 
 - First release!